Move instance sync to different package
